[Enhancement] Digest updater now checks whether there are any changes
This is a followup for the recent update of the digest updater in which
we allowed it to continue in case of an error. Let's not allow that and
pro-actively try to avoid such error by checking whether there are
actually any changes to be commited to the repository. If not, simply
tell this to user and end gracefully.

Also, this fixes 2 typos.