RHOAIENG-15242: ref(manifests): add some newlines and indentation into the json strings with package version information
Previously, we used to have every json string on a single line, which meant we had long hard-to-read lines in our manifests.
Now there is some human-friendly indentation in the files so it should be easier to read.

This is to nave no impact on the data stored, it is a code formatting change only.